pyproject.toml
[tool.poetry]
name = "flaskter"
version = "0.1.0"
description = ""
authors = ["Artem Sitd  <tym3@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
gunicorn = "^22.0.0"
flask-restful = "^0.3.10"
flask-sqlalchemy = "^3.1.1"
flasgger = "^0.9.7.1"
pytest = "^8.2.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
